corrected1 classes3_2

+++ src/by/epam/lab/Purchase.java
@@ -49,7 +49,7 @@ public class Purchase implements Comparable<Purchase> {
 
     @Override
     public String toString() {
-        return number + ";" + percent + ";" + toByn((int) getCost()) + ";" + day;
+        return number + ";" + percent + ";" + toByn(getCost()) + ";" + day;
     }
 
     @Override
@@ -57,7 +57,7 @@ public class Purchase implements Comparable<Purchase> {
         return this.number - purchase.number;
     }
 
-    public double getCost() {
+    public int getCost() {
         return (int) Math.round(PRICE * number * (100 - percent) / 10000) * 100;
     }
